# openapi.yaml
openapi: 3.0.3
info:
  title: HomeFlip API
  description: Production-ready house flipping and rental property management API with advanced contractor verification
  version: 1.0.0
  contact:
    name: HomeFlip Support
    email: support@homeflip.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.homeflip.com/v1
    description: Production server
  - url: https://staging-api.homeflip.com/v1
    description: Staging server
  - url: http://localhost:5000/api
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /properties:
    get:
      tags: [Properties]
      summary: List properties
      description: Get paginated list of properties
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [ANALYZING, UNDER_CONTRACT, IN_PROGRESS, COMPLETED, SOLD, CANCELLED]
        - name: propertyType
          in: query
          schema:
            type: string
            enum: [SINGLE_FAMILY, CONDO, TOWNHOUSE, MULTI_FAMILY, COMMERCIAL, LAND, OTHER]
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProperties'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Properties]
      summary: Create property
      description: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePropertyRequest'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /properties/{id}:
    get:
      tags: [Properties]
      summary: Get property
      description: Get property by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Property details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          description: Property not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Properties]
      summary: Update property
      description: Update property by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePropertyRequest'
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          description: Property not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Properties]
      summary: Delete property
      description: Soft delete property by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Property deleted
        '404':
          description: Property not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Contractors
  /contractors:
    get:
      tags: [Contractors]
      summary: List contractors
      description: Get paginated list of contractors with scoring
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: trade
          in: query
          schema:
            type: string
            enum: [GENERAL, GC, PLUMBING, ELECTRICAL, HVAC, ROOFING, PAINTING, FLOORING, OTHER]
        - name: minScore
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: hasActiveLicense
          in: query
          schema:
            type: boolean
        - name: hasActiveInsurance
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of contractors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContractors'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contractors/{id}:
    get:
      tags: [Contractors]
      summary: Get contractor details
      description: Get detailed contractor information with scoring
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Contractor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractorDetails'
        '404':
          description: Contractor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Sort field and direction (e.g., 'name:asc', 'score:desc')
      required: false
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_]+:(asc|desc)$'
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Error message
            details:
              type: object
              description: Additional error details
            requestId:
              type: string
              description: Request ID for tracking
...
    Property:
      type: object
      required: [id, userId, street, city, state, zipCode, fullAddress, propertyType, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        fullAddress:
          type: string
        propertyType:
          type: string
          enum: [SINGLE_FAMILY, CONDO, TOWNHOUSE, MULTI_FAMILY, COMMERCIAL, LAND, OTHER]
        bedrooms:
          type: integer
          nullable: true
        bathrooms:
          type: integer
          nullable: true
        squareFeet:
          type: integer
          nullable: true
        lotSize:
          type: number
          nullable: true
        yearBuilt:
          type: integer
          nullable: true
        condition:
          type: string
          enum: [EXCELLENT, GOOD, FAIR, POOR, NEEDS_WORK]
        purchasePrice:
          type: number
          nullable: true
        estimatedRepairCosts:
          type: number
          nullable: true
        estimatedARV:
          type: number
          nullable: true
        estimatedProfit:
          type: number
          nullable: true
        estimatedROI:
          type: number
          nullable: true
        estimatedTimeline:
          type: integer
          nullable: true
        status:
          type: string
          enum: [ANALYZING, UNDER_CONTRACT, IN_PROGRESS, COMPLETED, SOLD, CANCELLED]
        notes:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        documents:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePropertyRequest:
      type: object
      required: [street, city, state, zipCode, propertyType]
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        propertyType:
          type: string
          enum: [SINGLE_FAMILY, CONDO, TOWNHOUSE, MULTI_FAMILY, COMMERCIAL, LAND, OTHER]
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        squareFeet:
          type: integer
        lotSize:
          type: number
        yearBuilt:
          type: integer
        condition:
          type: string
          enum: [EXCELLENT, GOOD, FAIR, POOR, NEEDS_WORK]
        purchasePrice:
          type: number
        estimatedRepairCosts:
          type: number
        estimatedARV:
          type: number
        notes:
          type: string
    UpdatePropertyRequest:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        propertyType:
          type: string
          enum: [SINGLE_FAMILY, CONDO, TOWNHOUSE, MULTI_FAMILY, COMMERCIAL, LAND, OTHER]
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        squareFeet:
          type: integer
        lotSize:
          type: number
        yearBuilt:
          type: integer
        condition:
          type: string
          enum: [EXCELLENT, GOOD, FAIR, POOR, NEEDS_WORK]
        purchasePrice:
          type: number
        estimatedRepairCosts:
          type: number
        estimatedARV:
          type: number
        status:
          type: string
          enum: [ANALYZING, UNDER_CONTRACT, IN_PROGRESS, COMPLETED, SOLD, CANCELLED]
        notes:
          type: string
    PaginatedProperties:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ContractorDetails:
      type: object
      required: [id, name, overallScore, overallGrade]
      properties:
        id:
          type: string
        name:
          type: string
        companyName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        trades:
          type: array
          items:
            type: string
        yearsInBusiness:
          type: integer
          nullable: true
        totalProjects:
          type: integer
        totalValue:
          type: number
          nullable: true
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
        overallGrade:
          type: string
          enum: [A, B, C, D, F]
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        insuranceScore:
          type: integer
          minimum: 0
          maximum: 100
        experienceScore:
          type: integer
          minimum: 0
          maximum: 100
        reviewCount:
          type: integer
        projectCount:
          type: integer
        permitCount:
          type: integer
        licenses:
          type: array
          items:
            $ref: '#/components/schemas/ContractorLicense'
        policies:
          type: array
          items:
            $ref: '#/components/schemas/InsurancePolicy'
        legalEvents:
          type: array
          items:
            $ref: '#/components/schemas/LegalEvent'
    ContractorLicense:
      type: object
      required: [id, number, state, status]
      properties:
        id:
          type: string
        number:
          type: string
        state:
          type: string
        boardName:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, EXPIRED, SUSPENDED, REVOKED, PENDING]
        adminVerified:
          type: boolean
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        expiresOn:
          type: string
          format: date-time
          nullable: true
        sourceUrl:
          type: string
          nullable: true
    InsurancePolicy:
      type: object
      required: [id, type, insurerName]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [GL, WC, AUTO, UMBRELLA, OTHER]
        insurerName:
          type: string
        policyNumber:
          type: string
          nullable: true
        coverageEachOccur:
          type: number
          nullable: true
        coverageAggregate:
          type: number
          nullable: true
        expiresOn:
          type: string
          format: date-time
          nullable: true
        brokerName:
          type: string
          nullable: true
        certificateUrl:
          type: string
          nullable: true
    LegalEvent:
      type: object
      required: [id, type, severity, title]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [LAWSUIT, VIOLATION, COMPLAINT, OSHA_VIOLATION, BBB_COMPLAINT, OTHER]
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        title:
          type: string
        description:
          type: string
          nullable: true
        filedOn:
          type: string
          format: date-time
          nullable: true
        resolvedOn:
          type: string
          format: date-time
          nullable: true
        sourceUrl:
          type: string
          nullable: true
    PaginatedContractors:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContractorDetails'
        pagination:
          $ref: '#/components/schemas/Pagination'
...